requirements.txt
asgiref==3.2.10
Django==3.1
djangorestframework==3.14.0
gunicorn==21.2.0
packaging==23.1
psycopg2==2.9.7
pytz==2023.3.post1
sqlparse==0.4.4
